+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.2 (unreleased)
 ----------------
 
+- Replaced npairs function with much faster npairs_3d. 
+
 - Halotools is now Python 3.x compatible
 
 0.1 (2016-03-13)